move can_handle_event to separate header file
this file will be base for other future traits
and possibly concepts in the future
