+++ include/fsmpp2/detail/traits.hpp
@@ -1,0 +1,22 @@
+#ifndef FSMPP_DETAIL_TRAITS_HPP
+#define FSMPP_DETAIL_TRAITS_HPP
+
+namespace fsmpp2::detail
+{
+
+template<class T, class E>
+class can_handle_event
+{
+    template<class U>
+    static auto test(int) -> decltype(std::declval<U>().handle(std::declval<E>()), std::true_type{});
+
+    template<class>
+    static std::false_type test(...);
+
+public:
+    static constexpr auto value = std::is_same_v<std::true_type, decltype(test<T>(0))>;
+};
+
+} // namespace fsmpp2::detail
+
+#endif // FSMPP_DETAIL_TRAITS_HPP
